+++ odfdom/src/main/java/org/odftoolkit/odfdom/dom/element/draw/DrawMarkerElement.java
@@ -161,7 +161,7 @@ public class DrawMarkerElement extends OdfElement {
     SvgViewBoxAttribute attr =
         (SvgViewBoxAttribute) getOdfAttribute(OdfDocumentNamespace.SVG, "viewBox");
     if (attr != null && !attr.getValue().isEmpty()) {
-      return Integer.valueOf(attr.intValue());
+      return attr.intValue();
     }
     return null;
   }
@@ -175,7 +175,7 @@ public class DrawMarkerElement extends OdfElement {
   public void setSvgViewBoxAttribute(Integer svgViewBoxValue) {
     SvgViewBoxAttribute attr = new SvgViewBoxAttribute((OdfFileDom) this.ownerDocument);
     setOdfAttribute(attr);
-    attr.setIntValue(svgViewBoxValue.intValue());
+    attr.setIntValue(svgViewBoxValue);
   }
 
   /**
