+++ odfdom/src/main/java/org/odftoolkit/odfdom/dom/element/draw/DrawPathElement.java
@@ -255,7 +255,7 @@ public class DrawPathElement extends DrawShapeElementBase {
     SvgViewBoxAttribute attr =
         (SvgViewBoxAttribute) getOdfAttribute(OdfDocumentNamespace.SVG, "viewBox");
     if (attr != null && !attr.getValue().isEmpty()) {
-      return Integer.valueOf(attr.intValue());
+      return attr.intValue();
     }
     return null;
   }
@@ -269,7 +269,7 @@ public class DrawPathElement extends DrawShapeElementBase {
   public void setSvgViewBoxAttribute(Integer svgViewBoxValue) {
     SvgViewBoxAttribute attr = new SvgViewBoxAttribute((OdfFileDom) this.ownerDocument);
     setOdfAttribute(attr);
-    attr.setIntValue(svgViewBoxValue.intValue());
+    attr.setIntValue(svgViewBoxValue);
   }
 
   /**
