+++ odfdom/src/main/java/org/odftoolkit/odfdom/dom/element/draw/DrawPolylineElement.java
@@ -256,7 +256,7 @@ public class DrawPolylineElement extends DrawShapeElementBase {
     SvgViewBoxAttribute attr =
         (SvgViewBoxAttribute) getOdfAttribute(OdfDocumentNamespace.SVG, "viewBox");
     if (attr != null && !attr.getValue().isEmpty()) {
-      return Integer.valueOf(attr.intValue());
+      return attr.intValue();
     }
     return null;
   }
@@ -270,7 +270,7 @@ public class DrawPolylineElement extends DrawShapeElementBase {
   public void setSvgViewBoxAttribute(Integer svgViewBoxValue) {
     SvgViewBoxAttribute attr = new SvgViewBoxAttribute((OdfFileDom) this.ownerDocument);
     setOdfAttribute(attr);
-    attr.setIntValue(svgViewBoxValue.intValue());
+    attr.setIntValue(svgViewBoxValue);
   }
 
   /**
