+++ odfdom/src/main/java/org/odftoolkit/odfdom/dom/element/table/TableFilterConditionElement.java
@@ -134,7 +134,7 @@ public class TableFilterConditionElement extends OdfElement {
     TableFieldNumberAttribute attr =
         (TableFieldNumberAttribute) getOdfAttribute(OdfDocumentNamespace.TABLE, "field-number");
     if (attr != null && !attr.getValue().isEmpty()) {
-      return Integer.valueOf(attr.intValue());
+      return attr.intValue();
     }
     return null;
   }
@@ -148,7 +148,7 @@ public class TableFilterConditionElement extends OdfElement {
   public void setTableFieldNumberAttribute(Integer tableFieldNumberValue) {
     TableFieldNumberAttribute attr = new TableFieldNumberAttribute((OdfFileDom) this.ownerDocument);
     setOdfAttribute(attr);
-    attr.setIntValue(tableFieldNumberValue.intValue());
+    attr.setIntValue(tableFieldNumberValue);
   }
 
   /**
